Moving Schematic related classes and methods out of the BaseEditor into a BaseSchematic subclass Fixes on the get_text_space() method on AscEditor
